requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
pytorch_lightning==2.0.0
torch==2.0.0
torchvision==0.15.1
